# ENH: Flight simulation speed up