## Pull request type

- [x] Code maintenance (refactoring, formatting, tests)

## Checklist

- [x] Tests for the changes have been added (if needed)
- [x] Docs have been reviewed and added / updated
- [x] Lint (`black rocketpy/ tests/`) has passed locally 
- [x] All tests (`pytest tests -m slow --runslow`) have passed locally
- [x] `CHANGELOG.md` has been updated (if relevant)

## Current behavior
Simulation is taking too long to run!

## New behavior
Read the commit history, one by one.

Some comments:
* The `get_value_opt` method can be twice as faster as the normal `get_value` method.
* After the implementation, I'm seeing a total simulation time of around 3s, while before we had a simulation time of ~4s. I will add my tests here soon enough.

## Breaking change
- [x] No

## Additional information
I am still running some tests within this branch.