@Gui-FernandesBR 2024-03-26
Ok, here is more detailed comment regarding the performance of flight simulation.

- Everything was only possible due to `cProfile` and `snakeviz`, amazing tools!
- Pieces of code that we still need to improve:
    - `Flight.u_dot_generalized`: this method is at least x2 slower than the traditional u_dot method, still present. This is hard to speedup tho. 
    - `Function.set_source`: it is calling the _check_user_input, which really slows down the whole Function.__init__
    - `Function._check_user_input`: I think it is time to discuss if we really need this method running as default with rocketpy. It is taking a significant time of simulation.
    - I would also suggest the discussion on the usage of `np.arg_sort()` method to sort the x_array in the Function class. This also consumes time in order to avoid problems in the Function data source.
    - `Function.__mul__`: I tried my best, but there is still a lot of `if` statements slowing this function.  I'm open to suggestions here.
    - `Function.get_value_opt`: It is already very optimized, but it is not enough. Now this is even more crucial for the Flight class, this method is responsible for ~8% of simulation time.
- Is there any numerical integration scheme faster than LSODA? It seems to be taking around 60% of the simulation time. Plus it cannot be parallelized. 
- I created a **simple** test to compare the simulation times before and after changes:

```txt
Before the improvements done in this PR:
========================================

All 20 simulations done, it took 95.40774 seconds
Mean time: 4.77039
Standard deviation: 0.71902

All 20 simulations done, it took 92.84953 seconds
Mean time: 4.64248
Standard deviation: 0.66002

After the improvements done in this PR:
=======================================

All 20 simulations done, it took 50.86210 seconds
Mean time: 2.54310
Standard deviation: 0.12359

All 20 simulations done, it took 50.47766 seconds
Mean time: 2.52388
Standard deviation: 0.10919
```

Finally, I think this one is ready for review now, @RocketPy-Team/code-owners