@MateusStano 2024-03-27
Incredible PR with really necessary changes. Just want to point out that most commits and the whole PR is not a Maintanence (MNT) type, but Enhancement. Most commits change the behavior of the classes somewhat

> * I created a **simple** test to compare the simulation times before and after changes:

Can you also check in this test if the simulation results are all returning the same values? There are a lot of small changes to the function class and I just want to make sure none of them have any unintended behavior.


> * `Function.set_source`: it is calling the _check_user_input, which really slows down the whole Function.**init**
> * `Function._check_user_input`: I think it is time to discuss if we need this method running as default with rocketpy. It is taking a significant time for simulation.
> * I would also suggest the discussion on the usage of the `np.arg_sort()` method to sort the x_array in the Function class. This also consumes time to avoid problems in the Function data source.
> * `Function.__mul__`: I tried my best, but there are still a lot of `if` statements slowing this function.  I'm open to suggestions here.

I think the `Function` class is actually misused by us. Currently, it serves two distinct purposes:

1. Used to facilitate the simulation by taking care of discretization, interpolation, and extrapolation (a purely mathematical tool)
2. Used for post-simulation data analysis (plots, prints, and so on)

The first needs to be as fast and minimal as possible, with it being used only internally in the code.
The second benefits from it being robust at the cost of speed, and it potentially interacts with the user.

If we want to truly speed the simulation up by changing the `Function` class, some architectural changes are needed. Currently, though, I'd say that we should be focusing the `Function` class on its internal use in the simulation, and maybe all these validation functions (and `_check_user_inputs`) might not be a good implementation in the current state