@Gui-FernandesBR 2024-04-19
Thank you for your solid review and contributions, @MateusStano .

After hours of debugging, I finally fixed the problems within the Flight class.
I believe this PR is ready for a review now. We should be merging it soon.

The Function and Flight classes were the main focus of this PR.
I see there's still an improvement opportunity in the Environment class, but let's leave it for another PR.
There are some new TODOs in the code, hope you don't mind it.