@Gui-FernandesBR 2024-04-29
> Pretty good!
> 
> Question: How much is the simulation sped up now?

Running 100 simulations with the Calisto rocket:

**Before this PR:**
* Mean: 5.24839 s
* Std deviation: 1.82163 s

**After this PR:**
* Mean: 1.45942 s
* Std deviation: 0.45128 s

Overall, code much faster now! 
I even added a new if statement to force the code to use the old `u_dot` method in case the rocket motor is a SolidMotor.
I remember that we agreed with this in a previous meeting.
@MateusStano can you approve it again please?

Tests are not passing due to a problem of hdf5 in python3.8 running on macOS. We will need to discuss it this week.