@MateusStano 2024-05-01
> I even added a new if statement to force the code to use the old `u_dot` method in case the rocket motor is a SolidMotor.

This might be breaking the tests btw